Update README.md
A IOS as well as Android on React-Native CLI(SDK=36.0.0). This Application contains Drawer Navigator as well as the bottom tab navigator.The home page allows you to see Mindhug Blogs as well as read them. Also you can see the past events and also register yourself into live events on the events page,from the menu page you can go to the Therapists and partnership page as well as can subscribe to the Mindhug Newsletter. For more information refer to the youtube link.